Add files via upload
As of right now, this gets the user's home timeline. This code is able to handle getting both user and home timeline, we just need to add an if statement to handle that. Remember, once we get our view we will need to remove the main methods from these files.